requests 🌐

+++ app/api/reservations/remove-users/route.ts
@@ -1,10 +1,24 @@
 import { query } from "@/lib/db";
+import protect from "@/lib/protect";
 import NewReservationMember from "@/templates/reservationUserEdit/template";
 import { NextResponse } from "next/server";
 
 export async function POST(req: Request) {
   try {
     const { reservation, users } = await req.json();
+    const isAuthorized = (await protect(
+      req.headers.get("Authorization")
+    )) as any;
+
+    if (!isAuthorized) {
+      return NextResponse.json(
+        {
+          success: false,
+          message: "Auth failed",
+        },
+        { status: 500 }
+      );
+    }
     const [] = await Promise.all([
       query({
         query: `DELETE FROM users_reservations WHERE reservationId = ? AND userId IN(${users.map(
